Move server back to root dir
